refactor: class and function component demo
