+++ examples/todos/src/components/TodoList/Function.tsx
@@ -1,0 +1,16 @@
+import store from '../../store';
+import TodoList from './TodoList';
+
+const { useModel, useModelEffectsLoading } = store;
+
+export default function({ title }) {
+  const [ state, dispatchers ] = useModel('todos');
+  const effectsLoading = useModelEffectsLoading('todos');
+  return TodoList(
+    {
+      state: { ...state, title, subTitle: 'Function Component' },
+      dispatchers,
+      effectsLoading,
+    },
+  );
+}
