+++ examples/todos/src/components/TodoList/TodoList.tsx
@@ -1,9 +1,6 @@
 import React from 'react';
-import store from '../store';
 
-const { useModel, useModelEffectsLoading } = store;
-
-export function TodoList({ state, dispatchers, effectsLoading }) {
+export default function({ state, dispatchers, effectsLoading }) {
   const { title, subTitle, dataSource } = state;
   const { toggle, remove } = dispatchers;
 
@@ -33,17 +30,5 @@ export function TodoList({ state, dispatchers, effectsLoading }) {
         ))}
       </ul>
     </div>
-  );
-}
-
-export default function({ title }) {
-  const [ state, dispatchers ] = useModel('todos');
-  const effectsLoading = useModelEffectsLoading('todos');
-  return TodoList(
-    {
-      state: { ...state, title, subTitle: 'Function Component' },
-      dispatchers,
-      effectsLoading,
-    },
   );
 }
