+++ examples/todos/src/components/TodoList/index.ts
@@ -1,0 +1,4 @@
+import FunctionComponent from './Function';
+// import ClassComponent from './Class';
+
+export default FunctionComponent;
